+++ src/shared-ui/ErrorBoundary.tsx
@@ -1,7 +1,13 @@
+import { IIconProps } from 'azure-devops-ui/Icon';
 import { ZeroData } from 'azure-devops-ui/ZeroData';
+import css from 'classnames';
 import { Component, ErrorInfo, ReactNode } from 'react';
 interface Props {
   children: ReactNode;
+  className?: string;
+  title?: string;
+  description?: string;
+  iconProps?: IIconProps;
 }
 
 interface State {
@@ -26,11 +32,13 @@ export class ErrorBoundary extends Component<Props, State> {
   public render() {
     if (this.state.hasError) {
       return (
-        <div className="flex-row flex-center">
+        <div className={css('flex-row flex-grow', this.props.className)}>
           <ZeroData
             imageAltText={''}
-            primaryText="Error"
-            secondaryText="An error occurred when showing this section"
+            className="flex-grow"
+            iconProps={this.props.iconProps}
+            primaryText={this.props.title || 'Error'}
+            secondaryText={this.props.description || 'An error occurred when showing this section'}
           />
         </div>
       );